Upgrade magmad image to bionic to allow for proxied github clones (#581)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/581

This diff updates the base image used for magmad to bionic. This is necessary
to allow for FeG updates to be proxied through the orc8r. Specifically, the libcurl
version included in Xenial is too old to be used with an HTTPS proxy. As such, when
we clone the github repo to extract to new docker-compose file, the clone fails.

Unfortunately, there are no suitable backports for this library,
so the only option seems to be to upgrade. Since we use python3.5 and python3.6 comes
by default on bionic, we must install from the python `deadsnakes` PPA.

Reviewed By: xjtian

Differential Revision: D16935073

fbshipit-source-id: d85e60af33bd935ffb3fff125c8907116f585bf1