Upgrade corejs to 3.0 for everything *except* relay-compiler
Summary:
We use `flat` which is in corejs 3, but not 2.  However, 2 is required
for relay-compiler to execute.  The way relay-compiler links with corejs
is really weird, so installing them together doesn't work well.  Because
of this, we should split relay to another package, install the old corejs
there, and use corej3 everywhere else.

See https://github.com/facebook/relay/issues/2701 for mention of the workaround

This change also adds the polyfills to jest

Reviewed By: tcirstea

Differential Revision: D16971058

fbshipit-source-id: 76f8a72863373d949dd198a4ca154dd32e3f5a93