Refactor generic partial gateway update/read to delegate all configurator loading to per-model functions
Summary:
since a gateway can depend on its associations, metadata, parent associations, it doesn't really make that much sense to load the magmad gateway model and pass that on as a param to its per model functions.
Now it takes in the networkID and gatewayID so that they can pull in what they need from the configurator.

Reviewed By: xjtian

Differential Revision: D16967565

fbshipit-source-id: e25d1e56939272e0a61ffba0042e0f50af69781d