Refactor generic partial gateway update/read to delegate all configurator loading to per-model functions (#584)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/584

since a gateway can depend on its associations, metadata, parent associations, it doesn't really make that much sense to load the magmad gateway model and pass that on as a param to its per model functions.
Now it takes in the networkID and gatewayID so that they can pull in what they need from the configurator.

Reviewed By: xjtian

Differential Revision: D16967565

fbshipit-source-id: fc765ac39165fbe3481f90b7d45541a64fb604ea