Created a packet tracer that reports table_id which caused packet drop in the pipeline
Summary:
Implemented packet tracer CLI/grpc service that reports the table_id which caused the packet to be dropped.

One important thing to note here is that ovs already supports some kind of packet-tracing with:
```
sudo ovs-appctl ofproto/trace gtp_br0 in_port=LOCAL,arp,reg5=1
```
It iterates through the tables directly without creating any traffic.
Our implementation differs in a way that we set a specific register to some value which indicates that this packet is a test-packet. We also install a new flow that matches this register every time a drop-flow is installed.

Reviewed By: ssanadhya

Differential Revision: D16835471

fbshipit-source-id: fc9d876bcf3af4fc2b3394d2aeadf695d4d063e5