Add redirect and restrict_service as possible final unit action
Summary:
**Summary**
Charging credit might come with Final Unit Action (FUA) indicating how to treat the session when it runs out of credit. There are three types of action: `TERMINATE`, `REDIRECT`, and `RESTRICT_ACCESS`. Currently, only `TERMINATE` is supported and is applied to sessions. `REDIRECT` and `RESTRICT_ACCESS` are ignored.
This diff keeps track of all kinds of FUA. There will be a stacked diff handling `REDIRECT`.

**Implementation**
1. Store final action in the credit object
2. Pass the final action to Local Enforcer when necessary

**What is affected**
We could see some debug logs saying the redirect or restrict_access action is not supported if that is the final unit action of an exhausted credit.

Reviewed By: emakeev

Differential Revision: D16930863

fbshipit-source-id: a3affd3e1550ea6f24fa643fbd119b635dca9f27