Retrieve redirect server information in Gy-CCA
Summary:
**Summary**
In the Gy-CCA message, when the `FinalUnitIndication` is `true` and `FinalUnitAction` is `REDIRECT`, the `RedirectServer` field would be populated with redirection related information. `session_proxy` in FeG should parse this piece of information and pass to the `session manager` in AGW.

**Implementation**
1. Add the `RedirectServer` field in the struct for parsing diameter message
2. Add the `RedirectServer` field in proto message
3. Implement the conversion from diameter message to proto message

**What is affected**
Nothing should be affected as session manager did not take this information.
Gy-CCA should still be parsed and converted to proto message successfully.

Reviewed By: emakeev

Differential Revision: D17049184

fbshipit-source-id: 7955e50266cdb3e5722f636199cf6133c573582d