Deploy metrics containers on separate pods
Summary:
When running `helm upgrade` the new metrics pod would fail since it needed to get a lock on the prometheus data directory. The old pod still had the lock on it and wouldn't terminate until the new one was running. This required manual intervention to terminate the old pod every time an upgrade was released.

Splitting up the 3 containers (prometheus, alertmanager, configmanager) into 3 pods will solve the problem since prometheus should only be upgraded when the version of the container is changed (which will be rare). Configmanager which is updated on every release can restart without trouble.

Since these containers have shared data, they all need to run on the same node. Using pod affinity we can assure that this happens.

Reviewed By: xjtian

Differential Revision: D16988999

fbshipit-source-id: 74657c181a8435202ffca335ffcbe411659f03b3