Parallel process families for prometheus cache scrape with a worker pool (#582)
Summary:
In scrape, we iterate over the families to convert them to a string one by one. This can be done concurrently. I implement a worker pool to ensure we don't run out of memory when there are too many buckets.

I can't seem to run the benchmark tests on my end, so I'm not sure if the worker pool should be larger or smaller.
Pull Request resolved: https://github.com/facebookincubator/magma/pull/582

Reviewed By: xjtian

Differential Revision: D17119369

Pulled By: Scott8440

fbshipit-source-id: e21597fee974cee563bdac17ef54f89648b72c19