Extend swx_proxy to implement service_health servicer (#600)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/600

In order for the FeG HA mechanism to work properly, it
needs to be able to Disable/Enable diameter connection creation.
This ability already exists for `s6a_proxy` and `session_proxy`.

This diff adds the service_health servicer to `swx_proxy` to
allow the gateway health service to function properly for a CWF
deployment.

Reviewed By: xjtian

Differential Revision: D17098569

fbshipit-source-id: 9d4157bc0d75105cffb5f7aa1c2018de56e2dba1