Adding proto definition to write MME state
Summary:
As part of Stateless MME, we plan to move the in-memory state of each ITTI task
to redis data store. This would ensure that UEs which are registered do not
experience a detach-then-attach behavior whenever the MME process restarts. This
stack of changes implements statelessness for the MME and NAS task. The feature
is work in progress and "use_stateless" should be set to false in mme.yml.

Reviewed By: talkhasib

Differential Revision: D16992983

fbshipit-source-id: 1c87190dbf951d58ff7b33b5f85f6cefbef57dd1