view firing alerts + page flow
Summary:
* view firing alerts
* navigation between firing alerts, edit alerts, and new alert.

I feel like the way I did the routing is a bit clunky. Also, currently when you go firing alerts -> new alert -> cancel, it goes to edit alerts page. It would be ideal for it to go back to whatever page came before new alert (in this case, firing alerts). But this requires the new alerts component to have context for which component called it, which requires nesting, and it seems silly to me to nest the new alerts route twice - once in Firing Alerts and once in Editing Alerts. Would love any ideas here!

Reviewed By: Scott8440

Differential Revision: D17144914

fbshipit-source-id: c63f695cdd669c597a1de3d45271d077797fd982