Add Sessiond functionality to restore pipelined flows
Summary:
Adds the necessary functions to restart pipelined and pass the current flows for every subscriber.

For pipelined setup to work we maintain an epoch in sessiond and when the epoch sent from pipelined is different from the one in sessiond we call pipelined setup. The setup function will combine all subscriber sessions(basically all the rules that need to be activated) and create a setup request for pipelined.

Epoch is made global for pipelined as in the next diff we'll be accessing it from the enforcement controller for sesiond pipelined setup verification.

Reviewed By: fishlinghu

Differential Revision: D17221428

fbshipit-source-id: f622f617ddc2b4e60e37a9c4ff2cdb2822f85b00