GUTI and ECGI state to proto conversion
Summary:
As part of Stateless MME, we plan to move the in-memory state of each ITTI task
to redis data store. This would ensure that UEs which are registered do not
experience a detach-then-attach behavior whenever the MME process restarts. This
stack of changes implements statelessness for the MME and NAS task. The feature
is work in progress and "use_stateless" should be set to false in mme.yml.

This specific change adds state to proto conversions for common structs like
GUTI and ECGI.

Reviewed By: fishlinghu

Differential Revision: D17332236

fbshipit-source-id: 5850f6fca298830cfeef201ab8c2b95001a5cb65