Fix WifiMetrics queries
Summary: The aggregate query that ran if no specific gateway was selected was formatted incorrectly so it never returned data. This fixes it and also allows for previously dropped tags to show in the legend if there are multiple unique tag values for that tag name on the same chart.

Reviewed By: fannycchen

Differential Revision: D17378881

fbshipit-source-id: 0b0417e377d9ad98804b8dae66ed9611b3349e0c