Refactor and clean up StateReporter on gateway (#642)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/642

## What's Changed?
- Moved checkin_manager's functionalities such as kick starting bootstrap/cert check when checkin fails into StateReporter
- Moved checkin_manager's kernel_versions_installed polling logic into GatewayStatus as its own class
- Moved out the grpc client recycle logic into its own class to clean up the logic a bit.
- Added some basic unit tests (Increasing test coverage by a bit :) )
- Deleted checkin-manager.py

## Note
I've kept some of the print statements such as "Checkin Success" and "Checkin Fail" since many people rely on those for indication of cloud<->gw connection and don't want to make a disruption to the correctness of the docs just yet.

After this is stable, I can remove the cloud checkind service and fully migrate to state based pipeline.

I've also modified some error logs to reduce the space it takes.

From...
{F210485203}
To...
{F210485230}

## What is affected?
- Magmad no longer reports to checkind on the cloud
- Magmad will print "Checkin Success!" and all other related prints if state reporting succeeds/fails.

Reviewed By: xjtian

Differential Revision: D17193502

fbshipit-source-id: 59d07f907f233392e10bd8a9228a332079d44ff8