Added fluent-bit to magma-dev
Summary: Added a fluent-bit role to the ansible in the magma-dev deployment. This added fluent-bit to the gateway vm, along with the appropriate config to send syslog and MME logs to the server side fluentd instance. This involved setting some rsyslog config as well, so syslog is available over a TCP port.

Reviewed By: xjtian

Differential Revision: D17549762

fbshipit-source-id: 0fa92cf7077a0a85c21e35382b5bd08a0f9dac64