Update relay to use V1 network types for network routing (#660)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/660

V0 access network types contained a field called `FegNetworkID` that was
used to determine which FeG network to send requests to. This caused the `feg_relay`
service to have to import models from the `cwf` and `lte` plugins, creating ugly import
cycles. Due to this, CWAG <-> FeG relay wasn't properly implemented.

This diff corrects this by updating the types used in the relay to the versions
created in V1. It also cleans up an old and unnecessary env variable (FEG HW ID).

Reviewed By: xjtian

Differential Revision: D17597027

fbshipit-source-id: 77de8fa7444472661a8c6bd23dc3d5f2b1489485