Add mac_addr conversion to sessiond (#665)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/665

Sessiond recevies Mac address from the AAA
server as a bytes field. For C++ a bytes field is
represented as a std::string. Despite this, each char
is still a byte, not a properly formatted UTF-8 char. Thus
when this field is sent to the pipelined service as a string.

The following error occurs:
  sessiond         | I0904 05:11:23.442327     1 CreditPool.cpp:132] Credit init failed for imsi 001010000000087 and charging key 1
  sessiond         | [libprotobuf ERROR google/protobuf/wire_format_lite.cc:534] String field 'magma.lte.UEMacFlowRequest.mac_addr' contains invalid UTF-8 data when serializing a protocol buffer. Use the 'bytes' type if you intend to send raw bytes.
  sessiond         | I0904 05:11:23.442504     1 LocalSessionManagerHandler.cpp:163] Successfully initialized new session in sessiond for subscriber 001010000000087
  sessiond         | I0904 05:11:23.444804    10 PipelinedClient.cpp:141] Could not add flow for subscriber with UE MAC�_>��: Exception deserializing request!

This diff fixes this bug by converting the mac_addr to format <hex>:<hex>:<hex>:<hex>:<hex>:<hex> (modeled off of Go's net.ParseMAC().String())

Reviewed By: amarpad

Differential Revision: D17185217

fbshipit-source-id: 6caf922d1b77faba0369654de06028101048d70e