Correctly identify services contributing to list of metas (#675)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/675

In D17193502 the behaviour of StateReporter was changed such that it was now comparing the names of services (which are required to provide metas) against the metas themselves. This caused StateReporter to think that the metas were not being provided, which prevented checkins on SoMA devices.

This is resolved by comparing service names against the actual services that we get the metas from.

Reviewed By: al8

Differential Revision: D17677777

fbshipit-source-id: 6585280ef22eef9538ae6bd6f48a9aae557a8261