Add support for traffic gen to UE Simulator
Summary:
This diff adds support to the UE Sim service to
allow for triggering traffic tests. This functionality is
implemented by creating a new RPC method to trigger the
traffic. The traffic is ultimately generated using `iperf3`.

Currently, the CallingStationID is set to the MAC ADDR of the
OVS bridge on the test VM. This allows traffic to flow through
the CWAG since the eth_src for iperf is the bridge mac.

In the future, we should store ue mac address and

Reviewed By: xjtian

Differential Revision: D17542869

fbshipit-source-id: be6c2b42ac74502bdea6b688aeffca29342782b9