QR code based mobile app configuration
Summary:
FBC mobile app needs to be configured to know which TGNMS instance and which keycloak instance to point to. It also may need to receive other config information in the future.

Currently this is done at build time. This will not work once the app is deployed to the app stores. This allows the mobile app to be built once and configured to work with different TGNMS software instances

Reviewed By: mahyali

Differential Revision: D17570498

fbshipit-source-id: b158ac1df50d987f244dfbf43eb47238f345111d