Modify blobstore to support version clobbering
Summary:
With the upcoming implementation of the gateway state replication
framework, the Orchestrator needs to support version cloberring, where
the version provided in the blob is stored, rather than just incrementing
the previously stored version by 1.

Since we consider Gateways to serve as the "master" of their own state, this
will allow the Orchestrator to remain in sync with gateway State. If the version is
not provided in the Blob, normal version incrementation will be done. Rather than
creating a seperate method for this, the `CreateOrUpdate` method is modified to support
this usecase. This puts the clients in charge of how version incrementation is to be done.
Since all production usages of blobstore don't supply version, this won't affect any
existing behavior.

Reviewed By: themarwhal

Differential Revision: D17709197

fbshipit-source-id: 23421b7c851357fe9065185bb4f3392b364880b2