Add SyncVersions RPC method to the State service (#690)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/690

This diff adds the SyncVersions RPC method to the Orchestrator State service.
This method takes in Key, Type, and Version and returns the the Key, Type, and Version
for any state that differs in either Version, or existence (i.e (K,T) doesn't exist).

This will allow gateway components to efficiently re-synchronize their state by first
gathering which states differ, then either reporting states that differ, or fetching
states that differ, dependening on which direction the re-sync is being performed.

Reviewed By: xjtian

Differential Revision: D17727004

fbshipit-source-id: 7d60dfe5253a38c479cea629adb165f412a7839c