Add metrics to snmp type devices \o/
Summary: For agent devices that gather device state via SNMP, also submit the data as timeseries data to prometheus. These are interface counters, such as byte and packet counts for all/unicast/broadcast/errors. Currently, the entire state path is used, making the metric name very very long and not as human friendly as it could be. While this is less than ideal visually, this makes it really clear in the backend where the data comes from. Adjustments should be made on the UI side to clean this up.

Reviewed By: Mokon

Differential Revision: D17507486

fbshipit-source-id: b66d065927ae39fc4e4adfcb5d68078058cdd497