Adding APIs to interact with MME state
Summary:
As part of Stateless MME, we plan to move the in-memory state of each ITTI task
to redis data store. The feature is work in progress and "use_stateless" should
be set to false in mme.yml.

This specific change adds the MmeNasStateManager class which provides an
interface for the MME task to interact with the in-memory and persistent state.

Reviewed By: amarpad

Differential Revision: D15420010

fbshipit-source-id: 6d5803e68efb16bc8ae35565e0f01602c3ceb4ca