Adding state converter class for MME NAS state
Summary:
As part of Stateless MME, we plan to move the in-memory state of each ITTI task
to redis data store. The feature is work in progress and "use_stateless" should
be set to false in mme.yml.

This change adds a MmeNasStateConverter class to encapsulate all
functions to serialize/deserialize MME and NAS state. This is a helper class
used by MmeNasStateManager. Functions to serialize/deserialize UE context are
included here. The functions to serialize/deserialize EMM context will be part
of future commits.

Reviewed By: amarpad

Differential Revision: D17485428

fbshipit-source-id: f039c2fa1b8212864c209aef062e63102e83c503