Add nil check for gateway status reporting and error logs
Summary: prod cloud's checkind service is failling due to a nil dereference which is not happening in staging of now. Not sure why this is happening since this code was not changed recently, but adding nil check and logs to avoid this in the future and investigate why this is happening.

Differential Revision: D17911087

fbshipit-source-id: 72f2146e4354b7f7266ac82a0ef1788be65bba68