Add nil check for gateway status reporting and error logs (#727)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/727

prod cloud's checkind service is failling due to a nil dereference which is not happening in staging of now. Not sure why this is happening since this code was not changed recently, but adding nil check and logs to avoid this in the future and investigate why this is happening.

Reviewed By: ardzoht

Differential Revision: D17911087

fbshipit-source-id: bda4c8be343660a7926bbd5d6a478340a1eb5942