add test for ErrorQueue (#734)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/734

Changed ErrorQueue to `emplace_back` and `pop_front` instead of `emplace_front` and `pop_back` so that when using "get" the items are returned in the order they were enqueued  (previously they were in reverse order).

Flipped comparison operator of `maxSize` and `errors.size()`.

Added a unit test for ErrorQueue.

Reviewed By: Mokon

Differential Revision: D17930426

fbshipit-source-id: 963c315ddd47ec66d7d6ed8ba5687d0d14b0b5b2