fix snmp hanging
Summary:
This commit fixes a problem where the snmp event thread was hanging and
thus preventing new snmp requests from being fullfilled. The issue stemed from
the fact that net-snmp without per session useage is not thread safe. We were
using it in two threads though. The epoll thread to make requests and the snmp
loop thread to process responses. This causes an issue where a list becomes
infinite in size and the function never returns.

To get around this I moved the snmp loop thread into the epoll thread loop which
I was already meaning to do. In the future we can just confine snmp requests/
responses to a single thread by using future.via and then we can use the other
cores for other channels.

Reviewed By: al8

Differential Revision: D17915832

fbshipit-source-id: 6f2cc11b29aaf469b41d9a47c2cd156ff1e700cd