make ErrorQueue synchronized (#737)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/737

Added a `folly:Synchronized` around the list providing storage for `ErrorQueue`, using a write lock in `add()` and a read lock in `get()`.

I also added a unit test which creates 3 threads which rapidly perform 10,000 adds to the ErrorQueue. The lack of a runtime error means it's threadsafe (since in 30,000 adds in a few milliseconds you'd likely see an error otherwise).

Reviewed By: Mokon

Differential Revision: D17935120

fbshipit-source-id: c667ab00ed3e98bcb25255407a2b022fd19bc029