make Diff.h generic (#758)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/758

I made the diff() function in Diff.h more generic - it'll now work on lists in addition to sets.

I used SFINAE and overloaded the diff() function. The efficient diffing algorithm requires a sorted collection. If a collection has a "sort" member function, it'll be sorted, and if it doesn't the diff() function checks to make sure that the collection is already sorted.

If a collection doesn't have a sort() function and is unsorted, diff() throws an exception.

This won't change the previous behavior of diff() the way we were using it (on a set), but now allows us to use it on a list as well.

Reviewed By: Mokon

Differential Revision: D18033800

fbshipit-source-id: e300991581bc55bc8d7d45c170132a4ed801058c