fix gateways full view bug
Summary:
as requested by alan.

for some reason it looks like the ?view=full endpoint returns a 500 when a network with a mesh (but no gateway) exists.

if gatewayTKs is nil but there's a mesh associated with the network, configurator.LoadEntities will still return NetworkEntities but of type mesh. if we specify the orc8r.MagmadGatewayType parameter, though, and pass in a nil gatewayTKs, loadedGateways will be appropriately empty

Reviewed By: themarwhal

Differential Revision: D18063199

fbshipit-source-id: ab0c91ba67d2c67cd46f599605c22d9c67e1a035