install script for ubuntu (#767)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/767

I tested installing symphony agent on an ENS Ubuntu 18 VM.

Calculating a UID for the snowflake can be done the same way as on CentOS, and now does so for Ubuntu.

The default docker package on ubuntu has a broken `docker login` command - it tries to use an X11 display even though ubuntu server doesn't have one. The script checks to see if the ubuntu version of the package is installed or the official docker version. If it's the ubuntu version, it asks the user to re-install from docker, and provides two links on how to do so.

Reviewed By: Mokon

Differential Revision: D18089973

fbshipit-source-id: c1512e5e325f10b2b0b59ad47f52b0de85cf6897