Implement gateway state replication service (#770)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/770

This diff serves to implement the gateway state replication service.
This service allows for replication of Redis state on the gateway, back
to the Orchestrator's state service. Going forward, this will allow for
access gateway high-availability and mobility.

The state replication service works by maintaining Redis clients for
each type of state to be replicated and an in-memory mapping for each piece of
state. By periodically comparing version stamps, the state replication service
can determine what state has changed, and send a request to the Orchestrator
with those states.

When the service first starts, it resyncs state with the Orchestrator, updating
it's in-memory mapping with the states that are already in sync.

Note: The state replication service currently only handles additions and modfications
of state. As of now, deletions will not propagate back to the Orchestrator.

Reviewed By: xjtian

Differential Revision: D17969688

fbshipit-source-id: a532bfe5da2941b052afc63778fc31c23f5fd8e4