Add ability to configure how much traffic to generate (#773)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/773

For running the cwag integration test to test usage monitoring, it is helpful to be able to configure the volume of traffic sent by the ue simulator.
It's not exact as some packets are dropped. (I think I might have to play around with the bit rate.) But I think it is good enough for now .

Reviewed By: koolzz

Differential Revision: D18070921

fbshipit-source-id: eb8ba2434266c04c476d49f80319206b7b1ddef4