Update state replication service to use RedisFlatDict (#780)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/780

This diff changes the container that the gateway state replication service
uses to load Redis state. The GSR service now use a RedisFlatDict and expects keys
to be in the format <id>:<type>. The type must have an associated Serde that's been
configured in the `state.yml` config file.

This change was made due to recent design discussions that determined the MME should
be able to key using IMSI:* and be able to fetch all state for a given subscriber.

Reviewed By: themarwhal

Differential Revision: D18124573

fbshipit-source-id: b61ccfaeaebab53900f55116afdf8acdf104b943