Update state replication service to use RedisFlatDict
Summary:
This diff changes the container that the gateway state replication service
uses to load Redis state. The GSR service now use a RedisFlatDict and expects keys
to be in the format <id>:<type>. The type must have an associated Serde that's been
configured in the `state.yml` config file.

This change was made due to recent design discussions that determined the MME should
be able to key using IMSI:* and be able to fetch all state for a given subscriber.

Differential Revision: D18124573

fbshipit-source-id: 782d3ce76e133d578810717861210135d9c49aa0