Cleaning up EMM context and adding function to populate callbacks for SMC
Summary:
When the SMC EMM procedure is read from redis data store, the callback handlers
need to be repopulated locally. This change adds the functions for this.

Further, the memory allocated by NasStateConverter for EMM context needs to be
freed at clean-up. This change adds a function for the same.

Both these functions will be used in future diffs.

Reviewed By: ardzoht

Differential Revision: D18127424

fbshipit-source-id: ba4e51d65286323543d1126ee39a1c64a3af344a