Stop sessiond from deactivating service when monitoring quota has exceeded (#794)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/794

## Sessiond Background
When a session is created, the session manager will maintain the state of the session in an instance of `SessionState`. A session maintains, along with other things, two credit pools: `UsageMonitoringCreditPool` and `ChargingCreditPool`.

`UsageMonitoringCreditPool` maintains credit granted by PCRF. This credit is used to send updates to the PCRF after rules marked with the specified monitoring key uses up credit. It is only used for reporting and does not determine whether a service is deactivated or not.

`ChargingCreditPool` maintains credit granted by OCS. Sessiond is responsible for cutting off the user service if this quota is used up.

Both of the pools use instances of `SessionCredit` to maintain data usage.

## Existing Issue
Every time a credit is updated, it checks whether a service should be discontinued. This logic did not correctly factor in whether it is a monitoring credit or charging credit. So when a monitoring quota has exceeded, it would incorrectly decide to terminate the service.

## The Fix For Now
I fixed this issue temporarily by adding a credit type field to specify what type of credit it is. There are ways to figure out whether it is monitoring or charging by looking at the final action struct, but I decided initializing `SessionCredit` with a type is more explicit. Maybe I will considering making `MonitoringCredit` and `ChargingCredit` that inherit from `SessionCredit`, but I am not sure if that's an overkill :).

Reviewed By: amarpad

Differential Revision: D18173914

fbshipit-source-id: bc3ab8353f27162f3ec6778c0a6cbe9a6a210566