devmand build script should not override existing docker-compose override
Summary:
This commit makes it so the devmand build script does not override
an existing docker-compose.override.yml. This has the consequence that if a user
changes env varibles they need to delete the override if they want it
regenerated but it seems much easier for them to just edit the override and
expect that it won't be overriden.

Reviewed By: joemin

Differential Revision: D18124239

fbshipit-source-id: f75c504baa8a4ee3ce1ba5f02a78ebc8a39af81c