fix bug in GetListGatewaysHandler
Summary:
tldr: get all gateways returns things that aren't gateways

=========================
because of how configurator.LoadEntities works, if we don't have any entities of type gatewayType (line 215), then entityTKs will be nil, which is great.

but then on line 233, when we pass that nil set of types and keys to LoadEntities, we get *all* of the entities in network nid because the type-key filter (entityTKs) is nil, which means entsByTK is not nil (which it should be). this causes things to show up when they shouldn't when listing gateways.

we can either forego the TK filter approach and just use type filters (which turns line 233 into two configurator.LoadEntities calls, one for orc8r.MagmaGatewayType and one for gatewayType), or we can short circuit it the whole method if the length of ids (line 216) is 0.

I prefer what's been done here because it really ensures that we will only have ents of type MagmadGatewayType or gatewayType.

Reviewed By: themarwhal

Differential Revision: D18203724

fbshipit-source-id: a4badcd363f4f4bba26123665d118bad16f0ff7b