Only decrement session metric when session is found
Summary: The call to `.Dec()` was made even if the session was not found in the table, creating a new metric series without apn or imsi labels that had a value of -1. Change the call to only be made if the session was found and apn/imsi could be retrieved.

Reviewed By: themarwhal

Differential Revision: D18312685

fbshipit-source-id: 83ac261f1dd298e7f62f0e096ceb1dffec6453dc