Fix deserialization of static policies from Redis to sessiond
Summary: Static policies have not been deserializing properly in sessiond. In policydb, the PolicyRule protobufs are stored within a RedisState protobuf before being serialized and stored in Redis. This revision adds that extra deserialization step to sessiond. Now, sessiond will deserialize the policy first as a RedisState, and then from the serialized_msg field, deserialize that into the PolicyRule.

Reviewed By: themarwhal

Differential Revision: D18323885

fbshipit-source-id: 7498fff8f14eacf80ec739b193998959fef31832