Add app futures for all ryu controllers (#762)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/762

Controllers need to call other controllers for performing flow logic specific to other tables(f.e ue alls arp table, MagmaController calls StartupFlows, etc) To handle this we set a future for all apps that is then set when the controller is initialized with the datapath. This approach is tested already and we use it across the unit test framework.

Reviewed By: amarpad

Differential Revision: D17678802

fbshipit-source-id: ac5285464d893840e2daedf3f261d500befe65ad