Fix feg policydb ser/de inconsistency with the lte ser/de (#822)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/822

There was a recent change in the way PolicyRules streamed down from orc8r are serialized to be stored into redis.
Instead of directly storing the proto serialized PolicyRule, it is now wrapped into a RedisState struct and then serialized again.
The change was not propagated to the feg side, so doing that here.

Reviewed By: mpgermano

Differential Revision: D18357641

fbshipit-source-id: 469546378597071dccebcf0f04d06462d5739f7c