Generalize Metrics Component
Summary: The Metrics component was built with `gatewayID` hardcoded in it, meaning it could only be used for segmenting metrics on gatewayIDs. This moves the specific filter logic out of the Metrics component so the caller now passes in that information, making it much more usable. Specifically this will be used for making graphs that filter on IMSI, APN, etc. for CWF.

Reviewed By: tcirstea

Differential Revision: D18280798

fbshipit-source-id: 9785c51582f9f42cb82194225f5b753f55bb5c75