Change FeG docker network mode to host mode
Summary:
This diff changes the FeG's network mode to `host`. This is
due to a failure condition that was causing `control_proxy` to crash
anytime another service was crashing. In the case of a service crashing
due to a bad config error, we got into a unrecoverable state, since
mconfig won't work unless control_proxy is running.

Host networking mode also simplifies the feg configuration. This allows
the mconfig to be the single source of truth for configuration, rather than
having to configure the same port in the `mconfig` as the `docker-compose.yml`.

Reviewed By: themarwhal

Differential Revision: D18440203

fbshipit-source-id: 76292ff0c251b3bda28851c68f12248190fb1940