Add support for rule deactivation in CCA (2/2) (#844)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/844

Our implementation of session_proxy / sessiond currently does not support "Charging-Rule-Remove" in CCA, specifically CCA-Update. When usage monitor quota is exceeded, PCRF should send a CCA-U that specifies which rules should be removed.

This diff covers the sessiond side of changes. When a usage monitor update is received, it will now process the list of ruleIDs given by `rules_to_remove`  and make the `deactivate_flows_for_rules` call to pipelined.

Note: I haven't figured out a good way to handle error if pipelined fails to deactivate the rules. Sessiond just prints that it failed to do so now.

Reviewed By: amarpad

Differential Revision: D18427384

fbshipit-source-id: 639ab07e944985f853d4fcfbac13cdfee4555a6c