Add support for rule deactivation in CCA (1/2)
Summary:
Our implementation of session_proxy / sessiond currently does not support "Charging-Rule-Remove" in CCA, specifically CCA-Update. When usage monitor quota is exceeded, PCRF should send a CCA-U that specifies which rules should be removed.
This diff is updates the `CCADiameterMessage` struct to include a field for `CCADiameterMessage` so that the AVP can be deserialized into the struct properly. I've also added some logic on the `UpdateSession` to translate the list of `RuleNames` and `BaseNames` into a list of rule IDs and propagate them to sessiond.

Reviewed By: amarpad

Differential Revision: D18427333

fbshipit-source-id: 5eff0cd39c49bc2e6869d5d5e9f243a5a08afb5b