Add env variable option to set a default FramedIpv4Addr
Summary:
This diff adds an environment variable that allows for always sending FramedIpv4
AVP in CCRs to the PCRF. If the Access Gateway doesn't provided one (in the case of CWF),
a default will be used. This default can either be configured as an env-var or will otherwise
be set to `10.10.10.10`.

Reviewed By: themarwhal

Differential Revision: D18463169

fbshipit-source-id: 2097ee5ac17da27519c596151ff96e4649a83374