Add env variable option to set a default FramedIpv4Addr (#855)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/855

This diff adds an environment variable that allows for always sending FramedIpv4
AVP in CCRs to the PCRF. If the Access Gateway doesn't provided one (in the case of CWF),
a default will be used. This default can either be configured as an env-var or will otherwise
be set to `10.10.10.10`.

Reviewed By: themarwhal

Differential Revision: D18463169

fbshipit-source-id: e5260b2bcb54c6d1368364f7186144a6fc79c75d