Refactor RedisFlatDict to only store one type
Summary:
Having the ability to store multiple types of state was complicating
usages of the RedisFlatDict and exposed clients to be more vulnerable to usage
mistakes (i.e. not formatting the key correctly, etc.). This diff refactors
`RedisFlatDict` to only support a single type of state. It also updates current
uses of this dictionary to work properly with this refactoring.

Differential Revision: D18367354

fbshipit-source-id: 47a7dcc0040cd9a57a4e51dfac298d4e678b49e3