docker compose network mode host (#843)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/843

Change default network mode of the agent docker container to "host". This gives the container more network permissions. This was necessary to enable ping over ipv6. This is not a security concern as any necessary firewall rules are configured on the host and will apply to the container as well.

Reviewed By: Mokon

Differential Revision: D18435415

fbshipit-source-id: 835a2d902512a1dd5dfa7ff7c6ed9ed32d855a0a